+++ .github/workflows/docker_build_stable.yml
@@ -1,15 +1,13 @@
 name: Docker Build Stable
 
-on:
-  release:
-    types: [published]
-  
-jobs:
-  build_publish_docker_container:
-    runs-on: ubuntu-latest
+# This workflow is now handled by build_push.yml
+# Keeping this file to avoid breaking references, but it does nothing
 
-    steps:         
-      - name: run docker build and publish script
-        run: |
-          curl -X POST -H "Accept: application/vnd.github.v3+json" -H "Authorization: token ${{ secrets.DEPLOY_PREVIEW_TOKEN }}" -d '{"ref":"main", "inputs":{"tachidesk_release_type": "stable"}}' https://api.github.com/repos/suwayomi/docker-tachidesk/actions/workflows/build_container_images.yml/dispatches
-          
+on:
+  workflow_dispatch:
+
+jobs:
+  noop:
+    runs-on: ubuntu-latest
+    steps:
+      - run: echo "Docker builds are now handled by build_push.yml"
