+++ server/src/main/kotlin/suwayomi/tachidesk/server/user/UserType.kt
@@ -1,15 +1,17 @@
 package suwayomi.tachidesk.server.user
 
+import com.auth0.jwt.exceptions.JWTVerificationException
+import io.github.oshai.kotlinlogging.KotlinLogging
 import io.javalin.http.Context
-import io.javalin.http.Cookie
 import io.javalin.http.Header
-import io.javalin.http.SameSite
 import io.javalin.websocket.WsConnectContext
 import suwayomi.tachidesk.global.impl.util.Jwt
 import suwayomi.tachidesk.graphql.types.AuthMode
 import suwayomi.tachidesk.server.JavalinSetup.Attribute
 import suwayomi.tachidesk.server.JavalinSetup.getAttribute
 import suwayomi.tachidesk.server.serverConfig
+
+private val logger = KotlinLogging.logger {}
 
 sealed class UserType {
     class Admin(
@@ -83,19 +85,16 @@ fun getUserFromContext(ctx: Context): UserType {
                         // Set the new access token as a cookie with long Max-Age
                         // (matches refresh token expiry so browser doesn't clear it)
                         ctx.cookie(
-                            Cookie(
-                                name = "suwayomi-server-token",
-                                value = newAccessToken,
-                                maxAge = serverConfig.jwtRefreshExpiry.value.inWholeSeconds.toInt(),
-                                sameSite = SameSite.STRICT,
-                                isHttpOnly = true,
-                                secure = ctx.scheme() == "https",
-                                path = "/",
+                            AuthCookieUtil.createAuthCookie(
+                                ctx,
+                                "suwayomi-server-token",
+                                newAccessToken,
+                                serverConfig.jwtRefreshExpiry.value.inWholeSeconds.toInt(),
                             ),
                         )
                         return UserType.Admin(1)
-                    } catch (_: Exception) {
-                        // Refresh token is also invalid, user stays as Visitor
+                    } catch (e: JWTVerificationException) {
+                        logger.debug(e) { "Refresh token invalid, user remains as Visitor" }
                     }
                 }
             }
@@ -121,6 +120,9 @@ fun getUserFromWsContext(ctx: WsConnectContext): UserType {
         }
 
         AuthMode.UI_LOGIN -> {
+            // Note: WebSocket connections cannot auto-refresh tokens because
+            // WsConnectContext doesn't support setting response cookies.
+            // Clients should ensure valid tokens before establishing WS connections.
             val authentication =
                 ctx.header(Header.AUTHORIZATION) ?: ctx.header("Sec-WebSocket-Protocol") ?: ctx.cookie("suwayomi-server-token")
             val token = authentication?.substringAfter("Bearer ") ?: ctx.queryParam("token")
